Added function to insert new waypoint into active route

+++ error.h
@@ -15,9 +15,9 @@ namespace vultron
 	private:
 		std::string _message;
 		std::string _function;
-		int _line;
+		size_t _line;
 	public:
-		error(std::string message, std::string function, int line) :_message(message), _function(function), _line(line){}
+		error(std::string message, std::string function, size_t line) :_message(message), _function(function), _line(line){}
 		std::string toString()
 		{
 			return "MESSAGE: " + _message + "\nFUNCTION: " + _function + ".\nLINE: " + std::to_string(_line);
